Restructure parser
Ref. #91
